Remove non-informative logging messages
In my project evaluation I got the review finding of "Store logging
messages in the strings.xml". I am not a fan of storing such a messages
in the strings.xml as logging messages are part of the development
process (at least for this app) and they will never be translated of
subject of the change of phone localization. I think storing them there
would only pollute the strings.xml as you lost the context where they
are used. It just makes the code harder to read and maintain.

Anyway, I have to do the changes if I want to accept my project with the
Udacity reviewer. When I reviewed the messages (which were mainly placed
by the original developer) I found that they are not too consistent and
informative. In order to make them usable you need a strategy what to
log and when. Without that you just dump a lot of information that you
can hardly use for debugging.

As I said the existing logging messages are placed by the previous
developer, not by me. Why not me? Because I have not had a good strategy
to place them, and because I used the Android Studio debugger to
triangle the potential issues with my code. Logging is enabled only in
the dev mode, so it does not affect the production quality.

Because of these reasons I decided to remove these statements from the
code. If the code were bigger and/or more complex, a good logging
strategy should be found and we should introduce better logging
according to that.